:art::poop: Some migration towards more uniform libraries
Problem:
- Libraries are being migrated away from `gtest` towrard `Catch2`.
- The migration away from the minimal test library to `gtest` is
incomplete.

Solution:
- There isn't one.

Notes:
Migrating to `Catch2` from `gtest` is being done on other libraries to
avoid the trash that google engineers dump into `gtest` and the
additional work that it takes to use the library because it is overly
intrusive.  However, there is now an issue with Catch2 and c++23 when
using gcc.

It seems that the only test framework that is not causing problems
here is the minimal one. Perhaps the direction is not correct.